Add authors to twitter share experience
