+++ src/components/social-buttons/social-buttons.tsx
@@ -1,90 +1,61 @@
 import { useEffect, useState } from 'react';
-import theme from '@mdb/flora/theme';
+
+import { Author } from '../../interfaces/author';
+
+import { circleStyles } from './styles';
+import CopyLink from './copy-link';
 
 interface SocialButtonsProps {
     heading: string;
     description: string;
     className?: string;
+    authors?: Author[];
+    language?: string;
 }
-
-const circleStyles = {
-    borderRadius: 'circle',
-    background: theme.colors.background.containerInverse,
-    cursor: 'pointer',
-    padding: '6px',
-};
-
-const CopyLink: React.FunctionComponent<{ url: string }> = ({ url }) => {
-    // A lot of the styling is taken from Flora's Tooltip, but Flora hard-codes the on hover functionality to their Tooltip.
-    const [tooltipShown, setToolTipShown] = useState(false);
-    return (
-        <div sx={{ ...circleStyles, position: 'relative' }}>
-            <a
-                onClick={() => {
-                    navigator.clipboard.writeText(url);
-                    setToolTipShown(true);
-                    setTimeout(() => setToolTipShown(false), 2000);
-                }}
-                title="Copy link"
-            >
-                <img
-                    src="/developer/link.svg"
-                    alt="Copy Link"
-                    sx={{ width: '12px', height: '12px', display: 'block' }}
-                />
-            </a>
-            {tooltipShown && (
-                <div
-                    sx={{
-                        zIndex: 999,
-                        position: 'absolute',
-                        top: 28,
-                        bottom: 0,
-                        left: -4,
-                        right: 0,
-                    }}
-                >
-                    <div
-                        sx={{
-                            width: 0,
-                            height: 0,
-                            borderLeft: '8px solid transparent',
-                            borderRight: '8px solid transparent',
-                            marginLeft: 'base',
-                            borderBottom: `8px solid ${theme.colors.background.containerInverse}`,
-                        }}
-                    />
-                    <div
-                        sx={{
-                            bg: 'background.containerInverse',
-                            color: 'text.inverse',
-                            borderRadius: 'tooltips',
-                            padding: ['inc10', null, null, 'inc20'],
-                            textAlign: 'left',
-                            fontSize: ['inc00', null, null, 'inc10'],
-                            lineHeight: ['inc10', null, null, 'inc20'],
-                            fontFamily: 'body',
-                            width: 'max-content',
-                            boxShadow: 'level01',
-                        }}
-                    >
-                        Link Copied!
-                    </div>
-                </div>
-            )}
-        </div>
-    );
-};
 
 const SocialButtons: React.FunctionComponent<SocialButtonsProps> = ({
     heading,
     description,
     className,
+    authors,
+    language,
 }) => {
+    console.log(authors);
     const [url, setUrl] = useState('');
     useEffect(() => {
         setUrl(window.location.href);
     }, []);
+
+    const twitterTextElements: string[] = [];
+
+    if (authors) {
+        const authorTags = authors
+            .filter(
+                ({ twitter }) =>
+                    twitter && !twitter.toLowerCase().endsWith('/mongodb')
+            )
+            .map(({ twitter }) => {
+                const splitUrl = twitter?.split('/') as string[];
+                const username = splitUrl[splitUrl?.length - 1];
+                return `@${username}`;
+            })
+            .join(' ');
+        twitterTextElements.push(authorTags);
+    }
+    if (language) {
+        let languageTag: string;
+        if (language === 'C++') {
+            languageTag = 'cplusplus';
+        } else if (language === 'C#') {
+            languageTag = 'csharp';
+        } else {
+            languageTag = language.toLowerCase().replace(' ', '_');
+        }
+        twitterTextElements.push(languageTag);
+    }
+
+    twitterTextElements.push(heading);
+    console.log(twitterTextElements);
 
     return (
         <div
@@ -113,7 +84,9 @@ const SocialButtons: React.FunctionComponent<SocialButtonsProps> = ({
                 sx={circleStyles}
                 target="_blank"
                 rel="noreferrer"
-                href={`https://twitter.com/intent/tweet?url=${url}&text=${heading}`}
+                href={`https://twitter.com/intent/tweet?url=${url}&text=${encodeURIComponent(
+                    twitterTextElements.join(' ')
+                )}`}
                 title="Share on Twitter"
             >
                 <img
