+++ src/components/social-buttons/styles.ts
@@ -1,0 +1,8 @@
+import theme from '@mdb/flora/theme';
+
+export const circleStyles = {
+    borderRadius: 'circle',
+    background: theme.colors.background.containerInverse,
+    cursor: 'pointer',
+    padding: '6px',
+};
