+++ src/pages/[...slug].tsx
@@ -279,6 +279,7 @@ const ContentPage: NextPage<ContentPageProps> = ({
                     <SocialButtons
                         description={parseUndefinedValue(description)}
                         heading={title}
+                        authors={authors}
                         sx={{
                             gridArea: 'social',
                             justifySelf: ['start', null, 'end'],
